Prevent UB when righ/left shifting large char values
If a symbol value like two 0xFF's come in( symbolInputValue = (symbolInputValue << 8) + value)you will end up with a signed 0xFFFF which is negative and gives Undefined Behaviour later on.  Plus it doesn't make sense since we are just storing up to two unsigned chars at a time to extract the 6 bit value.